Reorganized client file structure to convey cleaner architecture
