+++ CloudDrive.App/Views/LoginWindow.xaml.cs
@@ -6,7 +6,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 
 
-namespace CloudDrive.App
+namespace CloudDrive.App.Views
 {
     public partial class LoginWindow : Window
     {
